backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
passlib[bcrypt]
pydantic